basket totals is now long and code cleanup (#128)
* cleaning code

* removed bug with int overflow

* create order history before baskets

* cleaning code

+++ shopBackend/src/main/java/com/example/shopbackend/model/BasketDTO.java
@@ -9,7 +9,7 @@ import java.util.List;
 public class BasketDTO {
 
     private final Long basketId;
-    private final int totalCost;
+    private final Long totalCost;
     private final List<BasketProductDTO> products;
 
     public BasketDTO(List<OrderQty> products) {
@@ -20,7 +20,7 @@ public class BasketDTO {
             sum += ordQty.getProduct().getPrice() * ordQty.getQuantity();
             this.products.add(new BasketProductDTO(ordQty));
         }
-        this.totalCost = sum;
+        this.totalCost = (long) sum;
         this.basketId = products.getFirst().getOrder().getId();
     }
